Separate inference method in order to break django dependency
